Indent: line up sccp_wrapper_asterisk_channel_read struct entries. Externalize the enum declaration.
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@6582 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/pbx_impl/ast/ast.h
@@ -149,7 +149,7 @@ static const struct sccp_pbx_devicestate {
  */
 static const struct sccp_extension_state {
 	const char *const text;
-	uint16_t extension_state;
+	int extension_state;
 } sccp_extension_states[] = {
 	/* *INDENT-OFF* */
 	{"Extension Removed",					AST_EXTENSION_REMOVED},
